WIP: Treat multiple timepoints as separate tiles
This is an initial stab that is not fully working.

Sometimes, it is probably bad to treat each timepoint as its own tile.
So maybe we need a way to distinguish? Or maybe this whole approach is
untenable, and we should simply update Bio-Formats to "Do The Right
Thing" with respect to Prairie datasets?